Add files via upload
In this image, you can see that the ATAC-seq for both stages shows peaks in this region, indicating open chromatin. There are far stronger peaks in the ATAC bigwig data for the larvae which makes sense because the noncoding atac bed file shows a non coding region there, in comparison to the diapause bed file which does not. The exon track in this case shows nothing since exons are for coding mRNA.