Add files via upload
Both larve and diapause ATAC data shows that there is a similar number of hits. at approx 30150 there is a peak in the larve DNA showing gene expression. The loop indicates interaction between 30,100 and 30,350(possibly and enhansor element. The ncORC both have 2 areas of non-coding before and after the peak of hits for the ATAC data.